Split execute.go to slightly improve readability
